third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-srcset-1x.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="resources/currentSrcHelper.js"></script>
<script>

    addEventListener("load", function() {
        test(function() {
            assert_equals(document.getElementById("dpr_1x").clientWidth, 400);
            assert_equals(document.getElementById("dpr_1x").naturalWidth, 400);
            assert_equals(document.getElementById("dpr_1x").naturalHeight, 400);
            assert_equals(currentSrcFileNameNoParams("dpr_1x"), "image-set-1x.png");
            assert_equals(document.getElementById("dpr_2x").clientWidth, 400);
            assert_equals(document.getElementById("dpr_2x").naturalWidth, 400);
            assert_equals(document.getElementById("dpr_2x").naturalHeight, 400);
            assert_equals(currentSrcFileNameNoParams("dpr_2x"), "image-set-2x.png");
        }, "Test srcset x descriptors on a 1x screen");
    }, false);
</script>
<!-- 'sizes' has no impact on 'x' or 'src' resources -->
<img id="dpr_1x" src="" srcset="../../hidpi/resources/image-set-1x.png?3 1x">
<img id="dpr_2x" src="" srcset="../../hidpi/resources/image-set-2x.png?3 2x">
